requirements.txt
akshare==1.17.38
numpy>=1.24
pandas>=1.5
pydantic-settings>=2.6
python-dotenv>=1.1
schedule>=1.2
fastapi>=0.115
uvicorn[standard]>=0.30
swagger-ui-bundle>=0.0.9
